fix(review): preserve isolated capture module boundaries
