+++ src/autoskillit/hooks/AGENTS.md
@@ -25,7 +25,7 @@ Sub-packages: guards/ (see guards/AGENTS.md), formatters/ (see formatters/AGENTS
 | `_command_classification.py` | Shared stdlib-only command classification primitives for guard scripts (interpreter/wrapper detection, git command classification) |
 | `_capture_contract.py` | Shared stdlib-only shell-capture transport limits and identifier validation |
 | `_policy_event.py` | Typed policy-event formatter for hook provenance messages (stdlib-only) |
-| `_capture_authority.py` | Minimal stdlib-only project/root descriptor authority and lifecycle context factory shared by producer and cleanup hooks |
+| `_capture/` | Small stdlib-only authority and lifecycle primitive package shared by producer and cleanup owners |
 | `_capture_artifacts.py` | Stdlib-only descriptor-anchored artifact publication, command runner, policy reader, and replay emitter |
 | `_capture_lifecycle.py` | Stdlib-only root-bound capture lifecycle ledger, writer leases, crash recovery, and quarantine deletion |
 | `capture_lifecycle_hook.py` | Cleanup-only `SessionStart`: bounded lifecycle sweep for interactive and headless sessions |
