+++ src/autoskillit/hooks/_capture/__init__.py
@@ -1,0 +1,1 @@
+"""Shared stdlib-only shell-capture primitives."""
